refactor: 코드 전체 구조 변경
