+++ src/main/java/baseball/Application.java
@@ -1,10 +1,7 @@
 package baseball;
 
-        import baseball.game.Game;
-
 public class Application {
     public static void main(String[] args) {
-        Game game = new Game();
-        game.start();
+        Game.start();
     }
 }
