+++ src/main/java/baseball/Computer.java
@@ -1,4 +1,4 @@
-package baseball.computer;
+package baseball;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Computer {
-    public List<Integer> generate() {
+    public static List<Integer> generate() {
         List<Integer> computer = new ArrayList<>();
         while (computer.size() < 3) {
             int randomNumber = Randoms.pickNumberInRange(1, 9);
