+++ src/test/java/baseball/ComputerTest.java
@@ -1,4 +1,4 @@
-package baseball.computer;
+package baseball;
 
 import org.junit.jupiter.api.Test;
 
@@ -9,8 +9,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ComputerTest {
     @Test
     void 컴퓨터_랜덤숫자_생성() {
-        Computer computer = new Computer();
-        List<Integer> randomNum = computer.generate();
+        List<Integer> randomNum = Computer.generate();
 
         assertThat(3).isEqualTo(randomNum.size());
     }
